requirements.txt
psycopg2-binary==2.9.3
python-decouple==3.5
